move slack api service to common
