+++ packages/server/customer-os-common-module/interfaces/slack.go
@@ -4,11 +4,26 @@ import (
 	"context"
 
 	postgres_entity "github.com/customeros/customeros/packages/server/customer-os-postgres-repository/entity"
+
+	"github.com/customeros/customeros/packages/server/customer-os-common-module/utils"
 )
 
 type SlackService interface {
 	GetSlackChannels(ctx context.Context, tenant string) ([]*postgres_entity.SlackChannel, error)
-	GetPaginatedSlackChannels(ctx context.Context, tenant string, page, limit int) ([]*postgres_entity.SlackChannel, int64, error)
+	GetPaginatedSlackChannels(ctx context.Context, tenant string, page, limit int) (*utils.Pagination, error)
 	StoreSlackChannel(ctx context.Context, tenant, source, channelId, channelName string, organizationId *string) error
 	SendMessageFromBot(ctx context.Context, channel, blocks string) error
+	GetSlackSettings(ctx context.Context, tenant string) (*SlackSettingsResponse, error)
+	ListSlackChannelsWithBot(ctx context.Context, tenant string) ([]SlackChannelResponse, error)
+	JoinSlackChannelsWithBot(ctx context.Context, tenant, channelId string) error
+	LeaveSlackChannelsWithBot(ctx context.Context, tenant, channelId string) error
+}
+
+type SlackSettingsResponse struct {
+	SlackEnabled bool `json:"slackEnabled"`
+}
+
+type SlackChannelResponse struct {
+	ID   string `json:"id"`
+	Name string `json:"name"`
 }
